+++ spec/factories/external_sources.rb
@@ -1,8 +1,6 @@
-require_relative '../../app/models/carto/external_source'
-
 FactoryGirl.define do
 
-  factory :external_source, class: Carto::ExternalSource do
+  factory :external_source, class: 'Carto::ExternalSource' do
     association :visualization, factory: :carto_visualization, type: 'remote'
     import_url 'http://www.wadus.com'
     rows_counted 1